freebsd: Fix memory leak in freebsd.ports.find
In function read_entries, we have to keep pointers returned by getline
in an array to be able to free them even if the man page in FreeBSD says
they will be freed when hdestroy is called. It turns out that the man
page is incorrect after reading the source code and POSIX standard.

In function write_maps, we can reset neither line nor len variable
because it will cause getline to malloc a new buffer every time it is
called. If we don't modify the values of these two variables, getline
will use realloc when it needs more space and we only have to free it
after leaving the loop.

I know freebsd.ports.find is a short-lived program and leaking memory
won't cause problems, but I still think we should do memory management
properly to allow using valgrind to check the program.